Fix extremely weird jest issue
